Initial Commit
I only changed things that I absolutely had to change while keep similar
class structure and having all the current desired functionalties that
were previously not working. Would I have done more? Yes, probably, but
this isn't my project. The main reasons the bugs were happening is
because when you were checking for file permissions you were writing 0
into the file that you than proceed to read from using Save(). But since
every time 0 is wrote into it. I've removed all saving code from game as
it is just for managing the save data. Save is where the save data is
written and read. I thought that was what you were going for, at least.
But it was not working like that at all before this.